requirements.txt
mattermostdriver
requests
websocket-client
python-slugify